Working Motion Magic Turn
Drive Sequence still buggy, turns seemingly forever

+++ src/org/usfirst/frc/team2635/robot/model/MotionParameters.java
@@ -2,10 +2,10 @@ package org.usfirst.frc.team2635.robot.model;
 
 public class MotionParameters {
 	
-	public double rightVelocity;
-	public double leftVelocity;
-	public double rightAcceleration;
-	public double leftAcceleration;
+	public int rightVelocity;
+	public int leftVelocity;
+	public int rightAcceleration;
+	public int leftAcceleration;
 	public double rightWheelRotations;
 	public double leftWheelRotations;
